refactor: unify chat section and update icons
