+++ frontend/src/components/ContactSection.tsx
@@ -1,14 +1,13 @@
 import { useLanguage } from '../context/LanguageContext'
-import viberIcon from '../assets/icons/viber.svg'
-import whatsappIcon from '../assets/icons/whatsapp.svg'
-import telegramIcon from '../assets/icons/telegram.svg'
-import emailIcon from '../assets/icons/email.svg'
+import phoneIcon from '../assets/icons/phone-call.png'
+import mailIcon from '../assets/icons/mail.png'
+import viberIcon from '../assets/icons/viber.png'
+import whatsappIcon from '../assets/icons/whatsapp.png'
 
 export default function ContactSection() {
   const { t } = useLanguage()
   const phone = '+359 881 234 567'
   const email = 'info@bulstaff.com'
-  const telegram = 'bulstaff'
   return (
     <section id="messengers" className="py-10 bg-gray-100">
       <div className="container mx-auto px-6 text-center">
@@ -17,8 +16,20 @@ export default function ContactSection() {
         </h2>
         <div className="mx-auto grid max-w-md grid-cols-1 gap-4 sm:grid-cols-2 md:max-w-none md:grid-cols-4">
           <a
+            href={`tel:${phone.replace(/\s+/g, '')}`}
+            className="flex items-center gap-4 p-4 bg-white rounded-lg shadow-sm hover:shadow-lg hover:-translate-y-1 transition-all"
+          >
+            <img src={phoneIcon} alt="Phone" className="h-8 w-8" />
+            <div className="text-left">
+              <p className="font-sans text-base text-primary font-medium">
+                {t('messengers.phone')}
+              </p>
+              <p className="font-sans text-base text-primary">{phone}</p>
+            </div>
+          </a>
+          <a
             href={`viber://chat?number=${phone.replace(/\s+/g, '')}`}
-            className="flex items-center gap-4 p-4 bg-white rounded-lg shadow-sm hover:shadow-lg hover:-translate-y-1 transition-all duration-200"
+            className="flex items-center gap-4 p-4 bg-white rounded-lg shadow-sm hover:shadow-lg hover:-translate-y-1 transition-all"
           >
             <img src={viberIcon} alt="Viber" className="h-8 w-8" />
             <div className="text-left">
@@ -30,7 +41,7 @@ export default function ContactSection() {
           </a>
           <a
             href={`https://wa.me/${phone.replace(/\D/g, '')}`}
-            className="flex items-center gap-4 p-4 bg-white rounded-lg shadow-sm hover:shadow-lg hover:-translate-y-1 transition-all duration-200"
+            className="flex items-center gap-4 p-4 bg-white rounded-lg shadow-sm hover:shadow-lg hover:-translate-y-1 transition-all"
           >
             <img src={whatsappIcon} alt="WhatsApp" className="h-8 w-8" />
             <div className="text-left">
@@ -41,22 +52,10 @@ export default function ContactSection() {
             </div>
           </a>
           <a
-            href={`https://t.me/${telegram}`}
-            className="flex items-center gap-4 p-4 bg-white rounded-lg shadow-sm hover:shadow-lg hover:-translate-y-1 transition-all duration-200"
-          >
-            <img src={telegramIcon} alt="Telegram" className="h-8 w-8" />
-            <div className="text-left">
-              <p className="font-sans text-base text-primary font-medium">
-                {t('messengers.telegram')}
-              </p>
-              <p className="font-sans text-base text-primary">@{telegram}</p>
-            </div>
-          </a>
-          <a
             href={`mailto:${email}`}
-            className="flex items-center gap-4 p-4 bg-white rounded-lg shadow-sm hover:shadow-lg hover:-translate-y-1 transition-all duration-200"
+            className="flex items-center gap-4 p-4 bg-white rounded-lg shadow-sm hover:shadow-lg hover:-translate-y-1 transition-all"
           >
-            <img src={emailIcon} alt="Email" className="h-8 w-8" />
+            <img src={mailIcon} alt="Email" className="h-8 w-8" />
             <div className="text-left">
               <p className="font-sans text-base text-primary font-medium">
                 {t('messengers.email')}
