+++ frontend/src/i18n/en.json
@@ -97,9 +97,9 @@
 
   "messengers": {
     "title": "Chat with us",
+    "phone": "Phone",
     "viber": "Viber",
     "whatsapp": "WhatsApp",
-    "telegram": "Telegram",
     "email": "Email"
   }
 }
