+++ frontend/src/pages/Home.tsx
@@ -5,7 +5,6 @@ import Services from '../components/Services'
 import JobList from '../components/JobList'
 import Testimonials from '../components/Testimonials'
 import Contact from '../components/Contact'
-import ContactSection from '../components/ContactSection'
 import AnimatedSection from '../components/AnimatedSection'
 import ParallaxSection from '../components/ParallaxSection'
 
@@ -34,7 +33,6 @@ export default function Home() {
       <AnimatedSection id="contact">
         <Contact />
       </AnimatedSection>
-      <ContactSection />
     </>
   )
 }
